+++ CHANGELOG.md
@@ -5,11 +5,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.9.3] - 2023-04-06
+## [Unreleased]
 
 ### Changed
 
 - Use Rollup and ESBuild to create a minimal bundle to run. (Some dependencies remain externalized.)
+
+## [0.10.1] - 2023-04-13
+
+### Fixed
+
+- Bug where /stats list would throw an error if the user wasn't tracking anything
+
+## [0.10.0] - 2023-04-06
+
+### Added
+
+- A new `/stats` command for stats tracking and competition
 
 ## [0.9.2] - 2023-04-05
 
@@ -181,7 +193,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Development Environment for needed dependencies.
 
-[0.9.3]: https://github.com/BYU-CS-Discord/CSBot/compare/v0.9.2...v0.9.3
+[Unreleased]: https://github.com/BYU-CS-Discord/CSBot/compare/v0.10.1...HEAD
+[0.10.1]: https://github.com/BYU-CS-Discord/CSBot/compare/v0.10.0...v0.10.1
+[0.10.0]: https://github.com/BYU-CS-Discord/CSBot/compare/v0.9.2...v0.10.0
 [0.9.2]: https://github.com/BYU-CS-Discord/CSBot/compare/v0.9.1...v0.9.2
 [0.9.1]: https://github.com/BYU-CS-Discord/CSBot/compare/v0.9.0...v0.9.1
 [0.9.0]: https://github.com/BYU-CS-Discord/CSBot/compare/v0.8.3...v0.9.0